sync w/ thirdparty fork h2
h2@b9d5397bd751633f676b3164ebe03cb3c4534a75
  => h2@e38678b1e2c465965f3ce5ec8d3040458415b376

- feat: implement HTTP/2 informational responses support

  Add support for HTTP/2 informational responses (1xx status codes) including
  103 Early Hints. This enables servers to send preliminary headers before
  the final response, improving client performance through early resource
  discovery and connection establishment.

  Changes include:
  - extend client and server APIs to handle interim informational responses
  - update stream state management for 1xx responses
  - add test for interim informational response scenarios

  This implementation follows RFC 7540 and RFC 8297 specifications for
  HTTP/2 informational responses handling.

- perf: Reduce huffman decode table size

- fix: auto-release padding from DATA frame

  We did not used to include padding when returning flow control.
  This now automatically releases any padded length if a DATA frame with padding is received.

- fix: Unparent connection span

  The `Connection` struct contains a `tracing::Span` tracking its
  lifetime. This can be useful. The problem is that this span has as its
  parent the span within which the connection is created, meaning that it
  keeps that span alive (not closed) while it is alive. In practice that
  means, for example, instrumented functions that create connections and
  store those connections for reuse will not have their spans closed and
  exported when they end.

  This commit sets the `Connection`'s span's parent to `None` and gives it
  a "follows-from" relation to the span in which it was created instead.
  This shows the causal relationship between them, without tying the
  creating span's lifetime to the connection's lifetime.

+++ rama-http-types/src/proto/h2/hpack/huffman/table.rs
@@ -264,7 +264,7 @@ pub(super) static ENCODE_TABLE: [(usize, u64); 257] = [
 
 // (next-state, byte, flags)
 #[allow(clippy::large_const_arrays)]
-pub(super) static DECODE_TABLE: [[(usize, u8, u8); 16]; 256] = [
+pub(super) static DECODE_TABLE: [[(u8, u8, u8); 16]; 256] = [
     // 0
     [
         (4, 0, 0x00),
